backend/requirements.txt
sqlalchemy~=2.0.35
pgvector~=0.3.5
psycopg2-binary~=2.9.9
asyncpg~=0.30.0
alembic~=1.13.3
uvicorn~=0.31.1
fastapi~=0.115.0
python-dotenv~=1.0.1
aiohttp~=3.11.13
redis~=5.2.1
aioredis~=2.0.1
redis[hiredis]
python-dateutil~=2.9.0
